+++ database/abstract_client.py
